Update tenzing.py to handle missing ORCID
Trying to fix another source of duplicate entries on the contributors page. 

At the moment, if a contributor has their ORCID on one Tenzing but not another, they get duplicate entries. The modifications to the script are an attempt to make the grouping more flexible so name is used if ORCID isn't available for a given contributor. May need a lot more work, though!