Update tenzing.py - fix websites duplicates bug (#231)
Trying to get the code to group by name and then propagate the ORCID across contributions for a given contributor. This is an attempt to fix the problem where if a contributor has a an  an ORCID on one Tenzing but not another, they end up with two separate entries on the website.

The code worked on a test df, but I'm very much out of my depth here, so if it's no good, just let me know!

Co-authored-by: Flavio Azevedo <falafla@gmail.com>